+++ src/lib/string/strspn.c
@@ -3,7 +3,7 @@
 
 #pragma           extrn Cstrchr
 
-int strspn(char *s, char *set) {
+size_t strspn(char *s, char *set) {
 	int	i = 0;
 	if (s) {
 		while (*s && strchr(set, *s))
